pyproject.toml
[tool.poetry]
name = "fache"
version = "0.1.0"
description = "Fact checker for dialog generation"
authors = ["Chunpeng Ma <ma.chunpeng@megagon.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
hydra-core = "^1.3.2"
fugashi = "^1.2.1"
ipadic = "^1.0.0"
transformers = "^4.30.2"
pytorch-lightning = "1.9.0"
scikit-learn = "^1.3.0"
torch = "1.12.1"
loguru = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
